MS-76: move Confirmation folder
